generate_mapping: add new '-k' option to set primary key
Sometimes, especially when working with experimental database dumps
or table views, a given table may not have a primary ket set yet. In
such cases, the `generate_mapping` tool generates a mapping file that
needs manual intervention.

This commit adds a new flag '-k' that forces a given column name to
be treated as primary key.